Refactor: move Destination to destination_io.h and add CWDestination variant support.
