+++ src/destination_io.h
@@ -26,4 +26,39 @@ namespace Standard {
 
 } // End Standard namespace
 
+/**
+ * Wrapper class for every supported address
+ */
+struct Destination {
+public:
+    explicit Destination() {}
+    explicit Destination(const CTxDestination& _dest, bool _isP2CS) : dest(_dest), isP2CS(_isP2CS) {}
+
+    CWDestination dest{CNoDestination()};
+    bool isP2CS{false};
+
+    Destination& operator=(const Destination& from)
+    {
+        this->dest = from.dest;
+        this->isP2CS = from.isP2CS;
+        return *this;
+    }
+
+    // Returns the key ID if Destination is a transparent "regular" destination
+    const CKeyID* getKeyID()
+    {
+        const CTxDestination* regDest = Standard::GetTransparentDestination(dest);
+        return (regDest) ? boost::get<CKeyID>(regDest) : nullptr;
+    }
+
+    std::string ToString() const
+    {
+        if (!Standard::IsValidDestination(dest)) {
+            // Invalid address
+            return "";
+        }
+        return Standard::EncodeDestination(dest, isP2CS ? CChainParams::STAKING_ADDRESS : CChainParams::PUBKEY_ADDRESS);
+    }
+};
+
 #endif //DESTINATION_IO_H
