+++ src/key_io.h
@@ -38,32 +38,4 @@ namespace KeyIO {
 
 }
 
-/**
- * Wrapper class for every supported address
- */
-struct Destination {
-public:
-    explicit Destination() {}
-    explicit Destination(const CTxDestination& _dest, bool _isP2CS) : dest(_dest), isP2CS(_isP2CS) {}
-
-    CTxDestination dest{CNoDestination()};
-    bool isP2CS{false};
-
-    Destination& operator=(const Destination& from)
-    {
-        this->dest = from.dest;
-        this->isP2CS = from.isP2CS;
-        return *this;
-    }
-
-    std::string ToString() const
-    {
-        if (!IsValidDestination(dest)) {
-            // Invalid address
-            return "";
-        }
-        return EncodeDestination(dest, isP2CS ? CChainParams::STAKING_ADDRESS : CChainParams::PUBKEY_ADDRESS);
-    }
-};
-
 #endif //PIVX_KEY_IO_H
