+++ src/qt/walletmodel.h
@@ -14,7 +14,6 @@
 #include "interfaces/wallet.h"
 
 #include "key.h"
-#include "key_io.h"
 #include "operationresult.h"
 #include "support/allocators/zeroafterfree.h"
